16/06/2021

+++ models/cancelled_order.js
@@ -41,7 +41,7 @@ const CancelledOrder = sequelize.define("cancelled_Order", {
     allowNull: false,
   },
   tax: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.DOUBLE,
     allowNull: false,
   },
   total: {
